+++ examples/mps/main.rs
@@ -50,13 +50,13 @@ fn main() {
 
     let vertex_buffer = device.new_buffer_with_data(
         vertices.as_ptr() as *const c_void,
-        (vertex_stride * vertices.len()) as u64,
+        vertex_stride * vertices.len(),
         buffer_opts,
     );
 
     let index_buffer = device.new_buffer_with_data(
         indices.as_ptr() as *const c_void,
-        (mem::size_of::<u32>() * indices.len()) as u64,
+        mem::size_of::<u32>() * indices.len(),
         buffer_opts,
     );
 
@@ -65,7 +65,7 @@ fn main() {
         .expect("Failed to create acceleration structure");
 
     acceleration_structure.set_vertex_buffer(Some(&vertex_buffer));
-    acceleration_structure.set_vertex_stride(vertex_stride as u64);
+    acceleration_structure.set_vertex_stride(vertex_stride);
     acceleration_structure.set_index_buffer(Some(&index_buffer));
     acceleration_structure.set_index_type(mps::MPSDataType::UInt32);
     acceleration_structure.set_triangle_count(1);
@@ -75,9 +75,9 @@ fn main() {
     let ray_intersector =
         mps::RayIntersector::from_device(&device).expect("Failed to create ray intersector");
 
-    ray_intersector.set_ray_stride(mem::size_of::<Ray>() as u64);
+    ray_intersector.set_ray_stride(mem::size_of::<Ray>());
     ray_intersector.set_ray_data_type(mps::MPSRayDataType::OriginMinDistanceDirectionMaxDistance);
-    ray_intersector.set_intersection_stride(mem::size_of::<Intersection>() as u64);
+    ray_intersector.set_intersection_stride(mem::size_of::<Intersection>());
     ray_intersector.set_intersection_data_type(
         mps::MPSIntersectionDataType::DistancePrimitiveIndexCoordinates,
     );
@@ -85,12 +85,12 @@ fn main() {
     // Create a buffer to hold generated rays and intersection results
     let ray_count = 1024;
     let ray_buffer = device.new_buffer(
-        (mem::size_of::<Ray>() * ray_count) as u64,
+        mem::size_of::<Ray>() * ray_count,
         MTLResourceOptions::StorageModePrivate,
     );
 
     let intersection_buffer = device.new_buffer(
-        (mem::size_of::<Intersection>() * ray_count) as u64,
+        mem::size_of::<Intersection>() * ray_count,
         MTLResourceOptions::StorageModePrivate,
     );
 
@@ -120,7 +120,7 @@ fn main() {
         0,
         &intersection_buffer,
         0,
-        ray_count as u64,
+        ray_count,
         &acceleration_structure,
     );
